requirements.txt
asgiref==3.7.2
Django==5.0.2
django-autoslug==1.9.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
pillow==10.2.0
psycopg2==2.9.9
PyJWT==2.8.0
pyscopg2==66.0.2
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
